Добавляет компонент main-empty при отсутствии предложений
