+++ project/src/components/pages/main/main-empty/main-empty.jsx
@@ -1,0 +1,26 @@
+import React from 'react';
+import {useSelector} from 'react-redux';
+import {getCurrentCity} from '../../../../store/reducer/main/selectors';
+
+function MainEmpty() {
+  const currentCity = useSelector(getCurrentCity);
+
+  return (
+    <div className="cities">
+      <div className="cities__places-container cities__places-container--empty container">
+        <section className="cities__no-places">
+          <div className="cities__status-wrapper tabs__content">
+            <b className="cities__status">No places to stay available</b>
+            <p className="cities__status-description">
+              We could not find any property available at the moment in
+              {currentCity}
+            </p>
+          </div>
+        </section>
+        <div className="cities__right-section"/>
+      </div>
+    </div>
+  );
+}
+
+export default MainEmpty;
